backend/requirements.txt
httpx==0.24.1
pytest==7.4.2
pydantic==1.10.0
fastapi==0.100.0
uvicorn==0.23.0
requests==2.31.0
openai==1.55.3
jira==3.8.0
pyjwt==2.10.1
